finished color ColorPicker
