+++ client/src/app/editor/editor tabs/EditorTabs.tsx
@@ -5,6 +5,7 @@ import { CloseOutline, ColorPalette } from 'react-ionicons'
 import { Colorful } from '@uiw/react-color'
 import { AnimatePresence, motion } from 'framer-motion'
 import TabButton from './TabButton'
+import ColorPicker from './ColorPicker'
 
 interface EditorProps {
   color: string
@@ -14,28 +15,6 @@ interface EditorProps {
 const EditorTabs = ({ color, setColor }: EditorProps) => {
   const [activeTab, setActiveTab] = useState(0)
   const [hexInput, setHexInput] = useState(color)
-
-  const onHexChange = (input: string) => {
-    const splitInput = input.split('#')
-    if (splitInput[splitInput.length - 1].length > 6) return
-
-    console.log(input)
-    if (input === '#' || input === '') {
-      setHexInput('#')
-    }
-
-    const inputRegex = /^[0-9a-fA-F]+$/
-
-    if (inputRegex.test(splitInput[splitInput.length - 1])) {
-      setHexInput('#' + splitInput[splitInput.length - 1])
-    }
-
-    const regex = /^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/g
-
-    if (regex.test('#' + splitInput[splitInput.length - 1])) {
-      setColor('#' + splitInput[splitInput.length - 1])
-    }
-  }
 
   return (
     <div className="flex-col absolute top-[50%] left-6">
@@ -47,39 +26,13 @@ const EditorTabs = ({ color, setColor }: EditorProps) => {
             <ColorPalette color={'white'} height="2.5em" width="2.5em" />
           }
         />
-        <AnimatePresence mode="wait">
-          {activeTab === 1 && (
-            <motion.div
-              key="color picker"
-              initial={{ opacity: 0, x: -6, y: '-50%' }}
-              animate={{
-                opacity: 1,
-                x: 20,
-              }}
-              exit={{ x: 4, opacity: 0 }}
-              transition={{
-                duration: 0.2,
-              }}
-              className="absolute top-[50%] left-full"
-            >
-              <Colorful
-                disableAlpha
-                color={color}
-                onChange={(color) => {
-                  setHexInput(color.hex)
-                  setColor(color.hex)
-                }}
-              />
-              <input
-                title="hex input"
-                value={hexInput}
-                onChange={(e) => onHexChange(e.target.value)}
-                onClick={(e) => e.currentTarget.select()}
-                className="w-[200px] mt-2 text-left bg-black border-2 border-black rounded-md"
-              />
-            </motion.div>
-          )}
-        </AnimatePresence>
+        <ColorPicker
+          isVisible={activeTab}
+          color={color}
+          setColor={setColor}
+          hexInput={hexInput}
+          setHexInput={setHexInput}
+        />
       </div>
     </div>
   )
